test: move dhis-service-reporting integration tests into dhis-test-integration
