+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/datastatistics/DataStatisticsEventStoreTest.java
@@ -39,11 +39,11 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
-import org.hisp.dhis.IntegrationTestBase;
 import org.hisp.dhis.analytics.SortOrder;
 import org.hisp.dhis.dashboard.Dashboard;
 import org.hisp.dhis.dashboard.DashboardService;
 import org.hisp.dhis.setting.SystemSettingManager;
+import org.hisp.dhis.test.integration.IntegrationTestBase;
 import org.hisp.dhis.user.UserSettingService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
